Organized components and updated imports
