+++ Enlace-main/src/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./app";
-import index from "./index.css";
+import "./index.css";
 
 ReactDOM.render(
   <React.StrictMode>
